move goToVersions to the versions store and move goToDashboard to the cloud store
