+++ packages/editor-ui/src/stores/cloudPlan.store.ts
@@ -167,11 +167,34 @@ export const useCloudPlanStore = defineStore(STORES.CLOUD_PLAN, () => {
 		state.initialized = true;
 	};
 
+	const goToDashboardPage = async () => {
+		if (usersStore.userIsOwnerInCloudDeployment) {
+			const dashboardLink = await generateCloudDashboardAutoLoginLink({
+				redirectionPath: '/dashboard',
+			});
+
+			location.href = dashboardLink;
+		}
+
+		return;
+	};
+
+	const generateCloudDashboardAutoLoginLink = async (data: {
+		redirectionPath: string;
+	}) => {
+		const searchParams = new URLSearchParams();
+
+		const adminPanelHost = new URL(window.location.href).host.split('.').slice(1).join('.');
+		const { code } = await getAutoLoginCode();
+		const linkUrl = `https://${adminPanelHost}/login`;
+		searchParams.set('code', code);
+		searchParams.set('returnPath', data.redirectionPath);
+
+		return `${linkUrl}?${searchParams.toString()}`;
+	};
+
 	return {
 		state,
-		initialize,
-		getOwnerCurrentPlan,
-		getInstanceCurrentUsage,
 		usageLeft,
 		trialDaysLeft,
 		userIsTrialing,
@@ -179,6 +202,11 @@ export const useCloudPlanStore = defineStore(STORES.CLOUD_PLAN, () => {
 		currentUsageData,
 		trialExpired,
 		allExecutionsUsed,
+		goToDashboardPage,
+		generateCloudDashboardAutoLoginLink,
+		initialize,
+		getOwnerCurrentPlan,
+		getInstanceCurrentUsage,
 		reset,
 		checkForCloudPlanData,
 		fetchUserCloudAccount,
